Check

+++ tp5/src/main/java/Layer.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 public class Layer {
     private int _n_neurons, _prev_n_neurons;
     private ArrayList<Neuron> _neurons;
-    private float _outputs[];
+    private double _outputs[];
 
     public Layer(int prev_n_neurons, int n_neurons, java.util.Random rand) {
         // all the layers/neurons must use the same random number generator
@@ -12,15 +12,15 @@ public class Layer {
 
         // allocate everything
         _neurons = new ArrayList<Neuron>();
-        _outputs = new float[_n_neurons];
+        _outputs = new double[_n_neurons];
 
         for (int i = 0; i < _n_neurons; i++)
             _neurons.add(new Neuron(_prev_n_neurons, rand));
     }
 
     // add 1 in front of the out vector
-    public static float[] add_bias(float[] in) {
-        float out[] = new float[in.length + 1];
+    public static double[] add_bias(double[] in) {
+        double out[] = new double[in.length + 1];
         for (int i = 0; i < in.length; i++)
             out[i + 1] = in[i];
         out[0] = 1.0f;
@@ -28,8 +28,8 @@ public class Layer {
     }
 
     // compute the output of the layer
-    public float[] evaluate(float in[]) {
-        float inputs[];
+    public double[] evaluate(double in[]) {
+        double inputs[];
 
         // add an input (bias) if necessary
         if (in.length != getWeights(0).length)
@@ -44,15 +44,15 @@ public class Layer {
             _outputs[i] = _neurons.get(i).activate(inputs);
 
         // bias treatment
-        _outputs[0] = 1.0f;
+        _outputs[0] = 1.0;
 
         return _outputs;
     }
 
     public int size() { return _n_neurons; }
-    public float getOutput(int i) { return _outputs[i]; }
-    public float getActivationDerivative(int i) { return _neurons.get(i).getActivationDerivative(); }
-    public float[] getWeights(int i) { return _neurons.get(i).getSynapticWeights(); }
-    public float getWeight(int i, int j) { return _neurons.get(i).getSynapticWeight(j); }
-    public void setWeight(int i, int j, float v) { _neurons.get(i).setSynapticWeight(j, v); }
+    public double getOutput(int i) { return _outputs[i]; }
+    public double getActivationDerivative(int i) { return _neurons.get(i).getActivationDerivative(); }
+    public double[] getWeights(int i) { return _neurons.get(i).getSynapticWeights(); }
+    public double getWeight(int i, int j) { return _neurons.get(i).getSynapticWeight(j); }
+    public void setWeight(int i, int j, double v) { _neurons.get(i).setSynapticWeight(j, v); }
 }
